+++ packages/react/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.16.1](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.16.0...@lubycon/react@1.16.1) (2021-08-22)
+
+**Note:** Version bump only for package @lubycon/react
+
+
+
+
+
+# [1.16.0](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.15.2...@lubycon/react@1.16.0) (2021-08-22)
+
+
+### Features
+
+* **react:** useCombinedRefs, useScrollEvent 훅 추가 ([#44](https://github.com/Lubycon/lubycon-frontend-libraries/issues/44)) ([d2ac59a](https://github.com/Lubycon/lubycon-frontend-libraries/commit/d2ac59a4569809c1b284f13b7b402d205165c710))
+
+
+
+
+
 ## [1.15.2](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.15.1...@lubycon/react@1.15.2) (2021-08-22)
 
 **Note:** Version bump only for package @lubycon/react